refactor(store): remove legacy space (#7559)
Closes: BS-136
