+++ packages/framework/store/src/store/index.ts
@@ -2,6 +2,6 @@ export type * from './collection.js';
 export { DocCollection } from './collection.js';
 export type * from './doc/block-collection.js';
 export * from './doc/index.js';
+export * from './id.js';
 export type * from './meta.js';
-export * from './space.js';
 export * from './store.js';
